requirements/base.txt
# GrowUp IoT System - Base Requirements
# Core dependencies needed for all environments

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# HTTP & Networking
requests==2.31.0
urllib3==2.1.0
websockets==12.0

# MQTT Client (for device control)
paho-mqtt==1.6.1

# Task Scheduling
APScheduler==3.10.4
pytz==2023.3

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
python-dateutil==2.8.2

# Logging
colorlog==6.8.0

# Data Processing
numpy==1.24.3
Pillow==10.1.0

# Hardware Interface (platform-independent)
smbus2==0.4.3
